requirements.txt
telethon
openai
gspread
oauth2client
python-dotenv
dropbox
json